requirements.txt
fastapi
uvicorn
python-dotenv
flask-cors
google-generativeai
